requirements.txt
black==22.12.0
Pillow==9.4.0
